+++ src/Generics/Deriving/TH.hs
@@ -342,7 +342,7 @@ deriveInstCommon genericName repName gClass fromName toName opts n = do
   -- See Note [Forcing buildTypeInstance]
   !(origTy, origKind) <- buildTypeInstance gClass useKindSigs name instTys
   tyInsRHS <- if repOptions opts == InlineRep
-                 then makeRepInline gt dv name cons origTy
+                 then repType gt dv name mempty cons
                  else makeRepTySynApp gClass dv name origTy
 
   let origSigTy = if useKindSigs